long-running-agent/requirements.txt
# Claude Agent SDK - Core dependency
claude-agent-sdk>=0.1.0

# Environment variables
python-dotenv>=1.0.0

# CLI argument parsing (built-in, but listed for clarity)
# argparse is part of Python standard library

# Async support
anyio>=4.0.0
